refactor: move onboarding to components
